Chunked upload fixed and tested

+++ zxs.php
@@ -1,6 +1,6 @@
 <?php
 /*
-	ZXS - simple web service for sharing files
+    ZXS - simple web service for sharing files
     Copyright (C) 2016 Dmitry V. Zimin
 
     This program is free software: you can redistribute it and/or modify
@@ -17,14 +17,12 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-// TODO: Delete partial uploaded files
-
 function delete_expired()
 {
-	$res = db_select("SELECT m.`id` FROM `zxs_files` AS m WHERE m.`deleted` = 0 AND m.`type` = 0 AND m.`expire` IS NOT NULL AND m.`expire` < CURDATE()");
-	foreach(@$res as $row)
+	$res = db_select("SELECT m.`id` FROM `zxs_files` AS m WHERE (m.`deleted` = 0 AND m.`type` = 0 AND m.`expire` IS NOT NULL AND m.`expire` < CURDATE()) OR (m.`deleted` = 1 AND m.`type` = 3 AND m.`date` IS NOT NULL AND DATE_ADD(m.`date`, INTERVAL 2 DAY) < CURDATE())");
+	if($res !== FALSE) foreach($res as $row)
 	{
-		$query = rpv_v2("UPDATE `zxs_files` SET `deleted` = 1 WHERE `id` = # LIMIT 1", array($row[0]));
+		$query = rpv_v2("UPDATE `zxs_files` SET `type` = 0, `deleted` = 1 WHERE `id` = # LIMIT 1", array($row[0]));
 		db_put($query);
 		unlink(UPLOAD_DIR."/f".$row[0]);
 	}
